requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
google-genai==1.70.0
python-dotenv==1.1.0
